bouton annuler et supprimer organisateur

+++ src/Controller/SortieController.php
@@ -127,10 +127,7 @@ class SortieController extends AbstractController
     #[Route('/{id}', name: 'app_sortie_delete', methods: ['POST'])]
     public function delete(Request $request, Sortie $sortie, EntityManagerInterface $entityManager): Response
     {
-        if(!$sortie->isIsPublished()) {
-
-        }
-        if ($sortie->getOrganisateur()->getEmail() === $this->getUser()->getUserIdentifier()) {
+        if ($sortie->getOrganisateur()->getEmail() === $this->getUser()->getUserIdentifier() && !$sortie->isIsPublished()) {
             if ($this->isCsrfTokenValid('delete' . $sortie->getId(), $request->request->get('_token'))) {
                 $entityManager->remove($sortie);
                 $entityManager->flush();
